Dernier update de la Bo Toutes les fonctionalités OK

+++ boV2/multipleDelete.php
@@ -64,24 +64,24 @@ if(isset($_POST['delete'])){
 
 	if (isset($_POST['id_team'])){
 
-        print_r($_POST);
-      print_r($_GET);
-
         $idToDelete = $_POST['delete'];
 
-        $isCreator = isCreatorOfTeam($_POST['id_team'], $_GET['userId']);
+        $isCreator = isCreatorOfTeam($_POST['id_team'], $user['id_utilisateur']);
 
         if (!empty($isCreator) || $_GET["userRole"] == 3){
 
-            $db = dbConnect();
+            if ($isCreator['createur'] != $user['id_utilisateur']){
 
-            foreach ($idToDelete as $value) {
+                $db = dbConnect();
 
-                $query = $db->prepare("DELETE FROM TEAMMATES WHERE id_team = :id_team AND id_user = :id_user");
-                $query->execute([
-                    "id_team" => $_POST['id_team'],
-                    "id_user"=>$value
-                ]);
+                foreach ($idToDelete as $value) {
+
+                    $query = $db->prepare("DELETE FROM TEAMMATES WHERE id_team = :id_team AND id_user = :id_user");
+                    $query->execute([
+                        "id_team" => $_POST['id_team'],
+                        "id_user"=>$value
+                    ]);
+                }
             }
         }
         header('Location: team.php?id='.$_POST["id_team"]);
